glHelper upd

+++ render/glhelper.cpp
@@ -3,8 +3,8 @@
 //float render::glHelper::m_zNear;
 //float render::glHelper::m_zFar;
 //float render::glHelper::m_fov;
-int render::glHelper::m_width;
-int render::glHelper::m_height;
+float render::glHelper::m_width;
+float render::glHelper::m_height;
 float render::glHelper::m_aspectratio;
 
 render::GLFrustum render::glHelper::m_viewFrustum;
